requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
bleach==6.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-cors-headers==4.0.0
django-dbbackup==4.0.2
django-filter==23.1
django-import-export==3.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.12
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.6
PyJWT==2.6.0
pymemcache==4.0.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
tzdata==2023.3
tzlocal==5.0.1
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
